Move TensorTranspose LibraryNode to STD.
