+++ dace/libraries/standard/environments/__init__.py
@@ -1,2 +1,3 @@
-# Copyright 2019-2021 ETH Zurich and the DaCe authors. All rights reserved.
+# Copyright 2019-2023 ETH Zurich and the DaCe authors. All rights reserved.
 from .cuda import CUDA
+from .hptt import HPTT
