--- dace/libraries/ttranspose/__init__.py
@@ -1,6 +1,0 @@
-# Copyright 2019-2023 ETH Zurich and the DaCe authors. All rights reserved.
-from dace.library import register_library
-from .nodes import *
-from .environments import *
-
-register_library(__name__, "ttranspose")
