--- dace/libraries/ttranspose/nodes/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2019-2023 ETH Zurich and the DaCe authors. All rights reserv
-from .ttranspose import TensorTranspose
